make prime_class.py
